requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
gTTS==2.5.4
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
pypiwin32==223
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
simple-websocket==1.1.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
